Refactoring publicPath and publicUrl methods
