+++ config/values.js
@@ -6,39 +6,13 @@
  */
 
 import * as EnvVars from './utils/envVars';
+import { getPublicUrl, getPublicPath } from './utils/publicPath';
 
 import codeSplittingConfigExtender from './extenders/codeSplitting';
 import singleRouteAppConfigExtender from './extenders/singleRouteApp';
 import reactApplicationExtender from './extenders/reactApplication';
 
-const host = EnvVars.string('HOST', 'localhost');
-const port = EnvVars.number('PORT', 3000);
-const publicUrl = EnvVars.string('PUBLIC_URL');
-const clientDevServerPort = EnvVars.number('CLIENT_DEV_PORT', 7331);
-const clientDevProxy = EnvVars.bool('CLIENT_DEV_PROXY', false);
-const NODE_ENV = EnvVars.string('NODE_ENV', 'development');
 const clientBundleWebPath = '/client/';
-
-function getPublicUrl() {
-  if (clientDevProxy && publicUrl) {
-    return `${publicUrl}/webpack`;
-  } else if (clientDevProxy) {
-    return `http://${host}:${port}/webpack`;
-  } else if (publicUrl) {
-    return publicUrl;
-  }
-  return `http://${host}:${clientDevServerPort}`;
-}
-
-function getPublicPath() {
-  if (NODE_ENV === 'development') {
-    // As we run a seperate development server for our client and server
-    // bundles we need to use an absolute http path for the public path.
-    return `${getPublicUrl()}${clientBundleWebPath}`;
-  }
-  // Otherwise we expect our bundled client to be served from this path.
-  return clientBundleWebPath;
-}
 
 const values = {
   // The configuration values that should be exposed to our client bundle.
@@ -66,19 +40,19 @@ const values = {
   // The public facing url of the app
   publicUrl: getPublicUrl(),
 
-  publicPath: getPublicPath(),
+  publicPath: getPublicPath(clientBundleWebPath),
 
   // The host on which the server should bind to.
-  host,
+  host: EnvVars.string('HOST', 'localhost'),
 
   // The port on which the server should bind to.
-  port,
+  port: EnvVars.number('PORT', 3000),
 
   // Should the webpack dev server be proxied through the public url
-  clientDevProxy,
+  clientDevProxy: EnvVars.number('CLIENT_DEV_PORT', 7331),
 
   // The port on which the client bundle development server should run.
-  clientDevServerPort,
+  clientDevServerPort: EnvVars.bool('CLIENT_DEV_PROXY', false),
 
   // This is an example environment variable which is used within the react
   // application to demonstrate the usage of environment variables across
